# P_Cloud_API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "P_Cloud_API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/authenticationTest": {
      "get": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ControlModule/{id}": {
      "get": {
        "tags": [
          "ControlModule"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ControlModule"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ControlModule"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ControlModuleInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ControlModule"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ControlModule": {
      "post": {
        "tags": [
          "ControlModule"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ControlModuleInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ControlModule"
            }
          }
        }
      }
    },
    "/api/ProcessData/{id}": {
      "get": {
        "tags": [
          "ProcessData"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "start",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "end",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProcessData"
              }
            }
          }
        }
      }
    },
    "/api/ProcessData": {
      "post": {
        "tags": [
          "ProcessData"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProcessData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Area": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "siteId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "site": {
          "$ref": "#/definitions/Site"
        },
        "processCells": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessCell"
          }
        }
      }
    },
    "ControlModule": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "equipmentModuleId": {
          "format": "int32",
          "type": "integer"
        },
        "superiorControlModuleId": {
          "format": "int32",
          "type": "integer"
        },
        "controlModuleInfoId": {
          "format": "int32",
          "type": "integer"
        },
        "controlModuleInfo": {
          "$ref": "#/definitions/ControlModuleInfo"
        },
        "equipmentModule": {
          "$ref": "#/definitions/EquipmentModule"
        },
        "superiorControlModule": {
          "$ref": "#/definitions/ControlModule"
        },
        "controlModuleInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModuleInfo"
          }
        },
        "inverseSuperiorControlModule": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModule"
          }
        },
        "processData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessData"
          }
        }
      }
    },
    "ControlModuleInfo": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "controlModuleId": {
          "format": "int32",
          "type": "integer"
        },
        "timestamp": {
          "format": "date-time",
          "type": "string"
        },
        "userIp": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "editTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "tolerance": {
          "format": "double",
          "type": "number"
        },
        "rangeLowerEnd": {
          "format": "double",
          "type": "number"
        },
        "rangeUpperEnd": {
          "format": "double",
          "type": "number"
        },
        "physicalUnitId": {
          "format": "int32",
          "type": "integer"
        },
        "controlModule": {
          "$ref": "#/definitions/ControlModule"
        },
        "editType": {
          "$ref": "#/definitions/EditType"
        },
        "physicalUnit": {
          "$ref": "#/definitions/PhysicalUnit"
        },
        "status": {
          "$ref": "#/definitions/Status"
        },
        "controlModules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModule"
          }
        }
      }
    },
    "EditType": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "controlModuleInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModuleInfo"
          }
        },
        "processData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessData"
          }
        }
      }
    },
    "Enterprise": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "sites": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Site"
          }
        }
      }
    },
    "EquipmentModule": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "unitId": {
          "format": "int32",
          "type": "integer"
        },
        "superiorEquipmentModuleId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "superiorEquipmentModule": {
          "$ref": "#/definitions/EquipmentModule"
        },
        "unit": {
          "$ref": "#/definitions/Unit"
        },
        "controlModules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModule"
          }
        },
        "inverseSuperiorEquipmentModule": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EquipmentModule"
          }
        }
      }
    },
    "PhysicalUnit": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "shortName": {
          "type": "string"
        },
        "fullname": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "controlModuleInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModuleInfo"
          }
        }
      }
    },
    "ProcessCell": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "areaId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "area": {
          "$ref": "#/definitions/Area"
        },
        "units": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Unit"
          }
        }
      }
    },
    "ProcessData": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "controlModuleId": {
          "format": "int32",
          "type": "integer"
        },
        "timestamp": {
          "format": "date-time",
          "type": "string"
        },
        "userIp": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "editTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "statusMessage": {
          "type": "string"
        },
        "error": {
          "type": "boolean"
        },
        "currentValue": {
          "format": "double",
          "type": "number"
        },
        "controlModule": {
          "$ref": "#/definitions/ControlModule"
        },
        "editType": {
          "$ref": "#/definitions/EditType"
        },
        "status": {
          "$ref": "#/definitions/Status"
        }
      }
    },
    "Site": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "enterpriseId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "enterprise": {
          "$ref": "#/definitions/Enterprise"
        },
        "areas": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Area"
          }
        }
      }
    },
    "Status": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "controlModuleInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ControlModuleInfo"
          }
        },
        "processData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProcessData"
          }
        }
      }
    },
    "Unit": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "processCellId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "processCell": {
          "$ref": "#/definitions/ProcessCell"
        },
        "equipmentModules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EquipmentModule"
          }
        }
      }
    }
  }
}